requirements.txt
# Core (required for all approaches)
anthropic>=0.40.0
openai>=1.0.0  # NVIDIA NIM uses OpenAI-compatible API
pandas>=2.0.0
openpyxl>=3.1.0

# Approach 2: Browser automation
playwright>=1.40.0
python-dotenv>=1.0.0

# Approach 3: Clipboard
pyperclip>=1.8.0
